requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Cloud SDKs
boto3==1.34.0
botocore==1.34.0
azure-mgmt-costmanagement==4.0.1
azure-mgmt-billing==6.0.0
azure-identity==1.15.0
google-cloud-billing==1.12.0
google-auth==2.25.2

# Data Processing
pandas==2.1.4
numpy==1.26.4
asyncio-mqtt==0.16.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# API & Web
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Indian Specific
razorpay==1.4.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Deployment
gunicorn==21.2.0
python-dotenv==1.0.0

# Postgres driver
psycopg2-binary==2.9.7

# bcrypt pinned for passlib compatibility
bcrypt==3.2.2
